+++ app/core/package.json
@@ -50,10 +50,12 @@
     "js-md5": "^0.7.3",
     "jsonfile": "^6.1.0",
     "lottie-web": "^5.11.0",
+    "mica-electron": "^1.4.0",
     "path-browserify": "^1.0.1",
     "pinia": "^2.1.3",
     "remixicon": "^2.5.0",
     "three": "^0.153.0",
+    "v-shared-element": "^3.1.1",
     "v-wave": "^1.5.1",
     "vanta": "^0.5.24",
     "vite-plugin-commonjs-externals": "^0.1.3",
