+++ app/core/renderer/components/addon/AgreementTemplate.vue
