+++ app/core/renderer/views/base/home/HomePage.vue
@@ -16,11 +16,8 @@
   </div>
 </template>
 
-<script name="Home" setup>
+<script name="HomePage" setup>
 
-onMounted(() => {
-  
-})
 </script>
 
 <style lang="scss" scoped>
