+++ app/core/renderer/views/layout/flat/FlatNavBar.vue
@@ -5,6 +5,7 @@
       <TouchMenuItem route="/home" name="Dashboard" icon="i-ri-home-3-line" />
       <TouchMenuItem route="/market" name="Market" icon="i-ri-quill-pen-line" />
       <TouchMenuItem route="/plugin" name="Plugin" icon="i-ri-plug-2-line" />
+      <TouchMenuItem route="/styles" name="Styles" icon="i-ri-paint-line" />
       <TouchMenuItem route="/setting" name="Setting" icon="i-ri-settings-6-line" />
       <p v-if="plugins.length" class="FlatNavBar-Title">PLUGINS</p>
       <TouchMenuItem @active="changeActivePlugin($event, item)" :doActive="() => activePlugin === item.name" v-for="item in plugins" :key="item.name">
